moved StaticType
