+++ src/lib.rs
@@ -225,64 +225,3 @@ impl Write for CallParameter {
         self.value.write(output)
     }
 }
-
-macro_rules! define_static_type {
-    ($($name:ident = $number:expr),*) => {
-        #[cfg_attr(feature = "serde", derive(serde::Serialize, serde::Deserialize))]
-        #[derive(Clone, Copy, Debug, Hash, PartialEq, Eq, PartialOrd, Ord)]
-        pub enum StaticType {
-            $($name = $number),*
-        }
-
-        impl TryFrom<i32> for StaticType {
-            type Error = ();
-
-            fn try_from(value: i32) -> Result<Self, Self::Error> {
-                match value {
-                    $($number => Ok(StaticType::$name),)*
-                    _ => Err(())
-                }
-            }
-        }
-
-        impl From<&StaticType> for i32 {
-            fn from(value: &StaticType) -> Self {
-                match value {
-                    $(StaticType::$name => $number,)*
-                }
-            }
-        }
-    };
-}
-
-define_static_type!(
-    Bool = 0,
-    Int = 1,
-    Float = 2,
-    String = 3,
-    Colour = 4,
-    Vector = 5,
-    Sound = 6,
-    Music = 7,
-    Object = 8,
-    ObjectSet = 9,
-    Transition = 10,
-    Easing = 11,
-    Sprite = 12,
-    Script = 13,
-    Layer = 14
-);
-
-impl Read for StaticType {
-    fn read(input: &mut impl std::io::Read) -> Result<Self, Error> {
-        let value = i32::read(input)?;
-
-        Self::try_from(value).map_err(|()| Error::InvalidStaticType(value))
-    }
-}
-
-impl Write for StaticType {
-    fn write(&self, output: &mut impl std::io::Write) -> Result<(), Error> {
-        i32::from(self).write(output)
-    }
-}
